back/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dateutil
python-jose[cryptography]
passlib
python-multipart
apscheduler
bcrypt
